purchase_form Done!

+++ src/main/java/com/example/CustomerInfo.java
@@ -5,11 +5,11 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class CustomerInfo {
     private Integer customerId;
-    private Integer total;
+    private Double total;
     @JsonFormat(pattern = "dd-MM-yyyy")
     private String date;
     
-    public CustomerInfo(@JsonProperty("customerId") Integer customerId, @JsonProperty("total") Integer total, @JsonProperty("date") String date) {
+    public CustomerInfo(@JsonProperty("customerId") Integer customerId, @JsonProperty("total") Double total, @JsonProperty("date") String date) {
         this.customerId = customerId;
         this.total = total;
         this.date = date;
@@ -23,11 +23,11 @@ public class CustomerInfo {
         this.customerId = customerId;
     }
 
-    public Integer getTotal() {
+    public Double getTotal() {
         return this.total;
     }
 
-    public void setTotal(Integer total) {
+    public void setTotal(Double total) {
         this.total = total;
     }
